新增转账手续费是挖矿激励的一部分

+++ helloworld-blockchain-core/src/main/java/com/xingkaichun/helloworldblockchain/core/utils/atomic/BlockChainCoreConstants.java
@@ -28,7 +28,7 @@ public class BlockChainCoreConstants {
     public final static long MINE_BLOCK_INCENTIVE_REDUCE_BY_HALF_INTERVAL_TIMESTAMP = 1 * 24 * 60 * 60 * 1000;
 
     //交易金额的小数点保留位数限制
-    public final static long TRANSACTION_AMOUNT_MAX_DECIMAL_PLACES = 8;
+    public final static int TRANSACTION_AMOUNT_MAX_DECIMAL_PLACES = 8;
     //最大交易金额
     public final static BigDecimal TRANSACTION_MAX_AMOUNT = new BigDecimal("1000000000000000000000000");
     //最小交易金额
